Fix String to double conversion

+++ src/main/java/geoview/calculators/MaintenanceScenariosCalculator.java
@@ -42,10 +42,10 @@ public class MaintenanceScenariosCalculator {
     
     //For now since we are treating all of these values as
     private static double calculateFeatureCost(String maintenanceType, Map<String, Object> feature){
-        double length = Integer.parseInt(feature.get("LENGTH").toString());
-        double depth = Integer.parseInt(feature.get("DEPTH").toString());
+        double length = Double.parseDouble(feature.get("LENGTH").toString());
+        double depth = Double.parseDouble(feature.get("DEPTH").toString());
         String pipeMaterial = feature.get("PIPE_MATERIAL").toString();
-        double radius = Integer.parseInt(feature.get("RADIUS").toString());
+        double radius = Double.parseDouble(feature.get("RADIUS").toString());
         
         double costPerFoot = length * getPipeTypeValue(pipeMaterial) * getReplacementTypeValue(maintenanceType) * getDepthWeight(depth) * getRadiusWeight(radius);
